part-4-telegram-bot/requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
Flask-Login==0.6.2
Flask-WTF==1.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.5.0
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
urllib3==2.0.3
Werkzeug==2.3.4
WTForms==3.0.1
